refactor : 토픽룸 참여자 목록 조회 API
